Extract popover top bar
